requirements.txt
django
pillow
django-cors-headers
requests
djangorestframework